refactor(vmm): move cpuid under x86_64
cpuid module is only relevant to x86_64 CPUs.
