+++ src/vmm/src/guest_config/templates/mod.rs
@@ -40,7 +40,7 @@ pub enum GetCpuTemplateError {
     #[cfg(target_arch = "x86_64")]
     /// Failed to get CPU vendor information.
     #[error("Failed to get CPU vendor information: {0}")]
-    GetCpuVendor(crate::guest_config::cpuid::common::GetCpuidError),
+    GetCpuVendor(crate::guest_config::x86_64::cpuid::common::GetCpuidError),
     /// CPU Vendor mismatched between the actual CPU and CPU template.
     #[error("CPU vendor mismatched between actual CPU and CPU template.")]
     CpuVendorMismatched,
